fix: Skip runtime operations in non-systemd environments
Cause: Both the actual role and the tests were previously assuming that
the system was booted with systemd and could run services. But this is
not the case when running the role during container builds.

Consequence: The role did not work during bootc container builds.

Fix: Detect if the system is booted (with systemd), and skip all runtime
operations and checks if not.

Result: The role now works during container builds.

The most tricky bit is the database initialization: `postgresql-setup`
calls `systemctl show` to query unit properties: Environment,
EnvironmentFiles, and NeedDaemonReload; but `systemctl show` requires a
running systemd pid 1. Work around that by replacing the systemctl
commands with `grep` on the unit file, and hardcode
`NeedDaemonReload=no`.

Setting the password (`postgresql_password`) and inputting an SQL file
(`postgresql_input_file`) both require a running server. Don't support
these options, at least for now. They *could* be supported by
temporarily starting `postmaster` during the role directly (not via
systemd). That works (I tested it manually), but is a bit finnicky.
Let's enable that in the future only if and when actually requested. In
the meantime, fail the role with clear error messages and adjust the
documentation.