fix: Disable postgresql_server_tuning for buildah connections
Cause: Detecting available RAM and basing the shared memory settings of
the server on that is futile and wrong during a container image build.
The available RAM during that build is independent of the actual
runtime, be that a bootc deployment or a system container.

Consequence: postgresql.service failed to start when the available RAM
during container build was higher than at deployment time. In the
opposite case, the shared memory tuning was too small and thus also
wrong.

Fix: Disable `postgresql_server_tuning` by default for container builds,
i.e. when being connected to a buildah instance.